requirements.txt
streamlit
pandas
transformers.AutoTokenizer, T5ForConditionalGeneration